### 2017-01-11 #157 @whedon
[PRE REVIEW]: bíogo/hts: high throughput sequence handling for the Go language
**Submitting author:** @kortschak (<a href="http://orcid.org/0000-0001-8295-2301">Robert Daniel Kortschak</a>)
**Repository:** <a href="https://github.com/biogo/hts" target ="_blank">https://github.com/biogo/hts</a>
**Version:** v1.0.0

Thanks for submitting your paper to JOSS kortschak. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
