### 2017-01-11 #157 @arfon
👋 @pjotrp - would you be willing to edit this submission for JOSS too? It's from the same authors as #154
