### 2017-01-14 e45e7c6
bgzf: update spec compliance warning
