### 2017-01-22 cfec6d9
handle case where b has fewer tags than a.
