+++ sam/sam_test.go
@@ -766,6 +766,16 @@ func (s *S) TestIssue32(c *check.C) {
 	c.Check(h.Progs()[0].UID(), check.Equals, "name")
 }
 
+func (s *S) TestEqualRefs(c *check.C) {
+	a, err := NewReference("aaa", "assem", "species", 1234, nil, nil)
+	c.Assert(err, check.IsNil)
+	b := a.Clone()
+	c.Assert(b.Set(Tag{'T', 'A'}, "xxx"), check.IsNil)
+
+	c.Assert(equalRefs(a, b), check.Equals, false)
+	c.Assert(equalRefs(b, a), check.Equals, false)
+}
+
 func BenchmarkParseCigar(b *testing.B) {
 	cig := []byte("69S17M5I30M1D45M1D23M5I14M2I4M1I10M2D7M1D6M14I33M1D6M1I7M1I18M1I8M1D4M1D4M2D57M1D21M1D6M1I14M1I7M1I3M1I9M1D3M1D7M1D37M1D9M1I5M1I15M4I12M1D10M1I10M1D8M1D26M7I12M1D20M1I36M1I22M3D8M1I23M1I13M2D10M1D12M1I15M6D4M1D4M1D1M2D5M1D3M17D1M1D13M3D7M1I29M2I9M1D2M4D7M2D8M5D3M1D1M1D23M1D10M6D19M3I24M1D8M1I11M6D14M1I5M8I12M1D8M2D5M2D2M1D23M1D11M4I35M2I19M1I4M1D13M7I33M1D21M3D2M1D9M4I19M1I14M1D7M1I41M1D23M3I18M1I6M1I13M1D9M1D1M1D20M1D23M5D8M1I13M2I11M1D78M2I18M10D9M2D10M1D10M2I6M1D3M1D21M2I7M1D7M2I12M1D20M2D18M1I12M1D8M4D18M1D6M1D20M1D14M1I1M2I23M1I10M1D7M1I15M1D4M1I9M1D11M1D12M1I8M1D21M1I13M2I59M1D12M1D18M1D13M1D22M1D13M1I19M1D13M1D19M1I11M2I27M2D10M1D17M6D13M2D17M1D13M1D19M1I3M1D13M2I33M1I26M2D9M2I21M2D10M1D36M1D32M5I23M1D13M2D17M1I14M2I24M1I5M2I8M2I24M2I9M1D7M1D2M1D15M3I19M1I2M1D3M1I7M1D5M2D24M5I1M4I33M1I13M3I34M1I2M1I23M1D3M2I8M1I5M5S")
 	for i := 0; i < b.N; i++ {
