### 2017-01-23 #157 @pjotrp
Dear Robert,

Thank you for submitting and sorry for the delay in review. I'll look for a reviewer. Meanwhile, can you look at some of the accepted papers to put in some more information - like proper citations? http://joss.theoj.org/papers/popular. Also a note on relevance, performance and explaining what SAM/BAM/CRAM is would be helpful to readers. Not asking you to do a lot of work, but take a cue from other papers.

Pjotr
