### 2017-01-23 #157 @kortschak
@pjotrp, thank you.

I am sort of at a loss to know how to address that. The SAM specification has no proper citation (no DOI as far as I know). Part of the issue here is that there are not good metrics to cover what we are aiming for here - reviewability of code without significant performance loss. Also note that we don't cover CRAM at this stage.

@brentp do you have any input on performance numbers? I know you've been making significant improvments there. So can you make comparisons under your workloads with samtools or other?

Dan
