### 2017-01-23 #157 @pjotrp
@kortschak I would check the samtools and sambamba papers - both bind to htslib

@brentp Brad has suggested you as a reviewer - would you be willing to take that on?
