### 2017-01-23 #157 @brentp
re performance (and utility) I just now wrote, in 30 minutes a simple program using biogo/hts to replicate the functionality in samtools flagstat:
https://gist.github.com/brentp/731895435be04f7d69ec69a6c2fcd008

On my machine, samtools flagstat takes 10.3 seconds.

the biogo/hts program takes 13.1 seconds. If I change (a single character) to allow it to use 2 cpus, then it take 8.4 seconds.
