### 2017-01-24 #157 @kortschak
@brentp If you would send a PR of that code to add it to biogo/hts/paper/examples or similar that would be good. Thanks.
