### 2017-01-24 97b2c43
add example replicating the functionality of samtools flagstat
