### 2017-01-24 #157 @arfon
@pjotrp - looks like we're good to open up the review issue at this point? If so then you'll need this command to create the review issue (with checklist etc):

```
> @whedon start review magic-word=bananas
```
