### 2017-01-24 biogo/hts#65 @sb10
README.md improvements
Hi, in reviewing for https://github.com/openjournals/joss-reviews/issues/168, I find that the "statement of need" requirement is not met by README.md. I think it would be satisfied if you just copied over the more comprehensive "Summary" paragraph from paper.md.

README.md also lacks example usage. I would be great to see a little trivial case example of using the API to do a basic operation like range through a (particular part of a) bam file and output to sam. Anything to get a friendly starting point in to using the API and exploring the methods further from there.

(The go docs for each package could also benefit from a friendly intro in a package comment, providing a starting point or method in to making use of that package.)

README.md may also be an appropriate place to include a brief sentence to satisfy the "Community guidelines" requirement of the joss review.
