### 2017-01-24 biogo/hts#66 @sb10
paper.md missing authors and references
Hi. In reviewing the paper for https://github.com/openjournals/joss-reviews/issues/168, I find that paper.md does not list authors and their affiliations. Could these be added?

There are also no DOI references. There is, for example, a published paper on the SAM spec out there.
