+++ paper/examples/flagstat/flagstat.go
@@ -1,27 +1,40 @@
+// Copyright ©2017 The bíogo Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
+// This package tabulates statistics on a bam file from the sam flag.
+// It replicates functionality in samtools flagstat.
 package main
 
 import (
 	"fmt"
 	"io"
-	"math"
+	"log"
 	"os"
 
 	"github.com/biogo/hts/bam"
 	"github.com/biogo/hts/sam"
 )
 
+const (
+	pass = iota
+	fail
+
+	mask = sam.Secondary | sam.ProperPair | sam.Supplementary | sam.Unmapped
+)
+
 func main() {
 	if len(os.Args) != 2 {
-		panic("Expecting a single bam argument")
+		log.Fatal("Expecting a single bam argument")
 	}
 	f, err := os.Open(os.Args[1])
 	if err != nil {
-		panic(err)
+		log.Fatal(err)
 	}
 	defer f.Close()
 	b, err := bam.NewReader(f, 0)
 	if err != nil {
-		panic(err)
+		log.Fatal(err)
 	}
 	defer b.Close()
 	b.Omit(2)
@@ -30,56 +43,68 @@ func main() {
 	var counts [2][12]uint64
 	// track mates on different chromosomes.
 	var mates [2][2]uint64
-	var fail int
+	var qc int
 	for {
 		read, err := b.Read()
 		if err == io.EOF {
 			break
 		}
 		if err != nil {
-			panic(err)
+			log.Fatal(err)
 		}
 		if read.Flags&sam.QCFail == 0 {
-			fail = 0
+			qc = pass
 		} else {
-			fail = 1
+			qc = fail
 		}
 
-		counts[fail][0]++
+		counts[qc][0]++
 		if read.Flags&sam.Supplementary != 0 {
-			counts[fail][11]++
+			counts[qc][11]++
 		} else if read.Flags&sam.Secondary != 0 {
-			counts[fail][log2(sam.Secondary)]++
+			counts[qc][Secondary]++
 		} else {
-			for i := 1; i < 12; i++ {
-				if read.Flags&(1<<uint(i)) != 0 {
-					counts[fail][i]++
+			for i := uint(1); i < 12; i++ {
+				if read.Flags&(1<<i) != 0 {
+					counts[qc][i]++
 				}
 			}
 		}
-		if read.Flags&(sam.Secondary|sam.ProperPair|sam.Supplementary|sam.Unmapped) == 0 {
+		if read.Flags&mask == 0 {
 			if read.MateRef != read.Ref && read.MateRef != nil && read.Ref != nil {
 				if read.MapQ > 4 {
-					mates[fail][1]++
+					mates[qc][1]++
 				}
-				mates[fail][0]++
+				mates[qc][0]++
 			}
 		}
 	}
 	// extract counts to match output from samtools flagstat.
-	fmt.Printf("%d + %d in total (QC-passed reads + QC-failed reads)\n", counts[0][0], counts[1][0])
-	fmt.Printf("%d + %d in total secondary\n", counts[0][log2(sam.Secondary)], counts[1][log2(sam.Secondary)])
-	fmt.Printf("%d + %d in total supplementary\n", counts[0][log2(sam.Supplementary)], counts[1][log2(sam.Supplementary)])
-	fmt.Printf("%d + %d duplicates\n", counts[0][log2(sam.Duplicate)], counts[1][log2(sam.Duplicate)])
[…]
