### 2017-01-25 f0b1ebb
add example replicating the functionality of samtools flagstat (#63)
* add example replicating the functionality of samtools flagstat
