### 2017-01-25 9a14d97
paper: extend summary to include citations (#64)
* samtool paper
* sambamba paper
