### 2017-01-25 biogo/hts#66 @kortschak
Hi @sb10, I am confused about the absence of authors and their affiliation. This is [present at the commit that was master](https://github.com/biogo/hts/blob/8e70c8e377b2b3c8b3aacf8b8180435a70ec6867/paper/paper.md) when I submitted the paper to JOSS. The current master retains that information.

A PR adding the Li 2009 and Tarasov 2015 papers has been merged.
