### 2017-01-25 a98f0f4
paper/examples/flagstat: use heavier benchmark and match logic
