+++ paper/examples/flagstat/flagstat.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// This package tabulates statistics on a bam file from the sam flag.
+// This program tabulates statistics on a bam file from the sam flag.
 // It replicates functionality in samtools flagstat.
 package main
 
@@ -13,6 +13,7 @@ import (
 	"os"
 
 	"github.com/biogo/hts/bam"
+	"github.com/biogo/hts/bgzf"
 	"github.com/biogo/hts/sam"
 )
 
@@ -30,17 +31,26 @@ func main() {
 		log.Fatal(err)
 	}
 	defer f.Close()
+	ok, err := bgzf.HasEOF(f)
+	if err != nil {
+		log.Fatal(err)
+	}
+	if !ok {
+		log.Println("EOF block missing")
+	}
+
 	b, err := bam.NewReader(f, 0)
 	if err != nil {
 		log.Fatal(err)
 	}
 	defer b.Close()
-	b.Omit(2)
+	b.Omit(bam.AllVariableLengthData)
 
 	// counts is indexed by [pass/fail][sam.Flag] where we have 12 possible sam Flags.
 	var counts [2][12]uint64
 	// track mates on different chromosomes.
-	var mates [2]struct{ low, high uint64 }
+	var mates [2]struct{ allMapQ, highMapQ uint64 }
+	var good, singletons, paired [2]uint64
 	var qc int
 	for {
 		read, err := b.Read()
@@ -56,55 +66,69 @@ func main() {
 			qc = fail
 		}
 
-		counts[qc][0]++
-		if read.Flags&sam.Supplementary != 0 {
-			counts[qc][Supplementary]++
-		} else if read.Flags&sam.Secondary != 0 {
-			counts[qc][Secondary]++
-		} else {
-			for i := uint(1); i < 12; i++ {
-				if read.Flags&(1<<i) != 0 {
-					counts[qc][i]++
-				}
+		for i := Paired; i <= Supplementary; i++ {
+			if read.Flags&(1<<i) != 0 {
+				counts[qc][i]++
 			}
 		}
 
-		const mask = sam.Secondary | sam.ProperPair | sam.Supplementary | sam.Unmapped
-		if read.Flags&mask == 0 {
+		const goodMask = sam.ProperPair | sam.Unmapped
+		if read.Flags&goodMask == sam.ProperPair {
+			good[qc]++
+		}
+
+		const mapMask = sam.MateUnmapped | sam.Unmapped
+		switch read.Flags & mapMask {
+		case sam.MateUnmapped:
+			singletons[qc]++
+		case 0:
+			paired[qc]++
 			if read.MateRef != read.Ref && read.MateRef != nil && read.Ref != nil {
 				if read.MapQ > 4 {
-					mates[qc].high++
+					mates[qc].highMapQ++
 				}
-				mates[qc].low++
+				mates[qc].allMapQ++
 			}
 		}
 	}
+
 	// extract counts to match output from samtools flagstat.
 	fmt.Printf("%d + %d in total (QC-passed reads + QC-failed reads)\n", counts[pass][Paired], counts[fail][Paired])
 	fmt.Printf("%d + %d in total secondary\n", counts[pass][Secondary], counts[fail][Secondary])
 	fmt.Printf("%d + %d in total supplementary\n", counts[pass][Supplementary], counts[fail][Supplementary])
 	fmt.Printf("%d + %d duplicates\n", counts[pass][Duplicate], counts[fail][Duplicate])
[…]
