### 2017-01-25 biogo/hts#66 @sb10
Yes, sorry, I mentally skipped over the table. The referenced PR will indeed fix the reference issue. (I'll be able to tick these items off in the review checklist once a new release tag is created with all the necessary changes.)
