### 2017-01-27 b71cda5
README: add summary, example and community guidelines
