### 2017-01-31 biogo/hts#65 @sb10
PS. Once the above PR is integrated along with the paper changes and a new release is made and tagged, I can close out the joss review.
