### 2017-02-01 #168 @sb10
v1.0.1 passes all of the reviewer questions and I'm satisfied with it.
