### 2017-02-02 14a9816
paper/examples/flagstat: add multithreaded samtools
