### 2017-02-04 #168 @arfon
@kortschak - could you please move the references in https://github.com/biogo/hts/blob/master/paper/paper.md to a separate `paper.bib` file and then reference them from your `paper.md` file. You can see an example of how to do this here: https://github.com/arfon/fidgit/tree/master/paper
