### 2017-02-04 521e3ec
paper: properly mark up references
