### 2017-02-04 a19f56e
paper: properly mark up references
