+++ paper/paper.bib
@@ -1,0 +1,30 @@
+@online{SAMspec,
+  author  = {The SAM/BAM Format Specification Working Group},
+  title   = {Sequence Alignment/Map Format Specification},
+  year    = {2016},
+  url     = {https://samtools.github.io/hts-specs/SAMv1.pdf},
+  urldate = {2017-04-2}
+}
+
+@Article{li,
+  author    = {Li H., Handsaker B., Wysoker A., Fennell T., Ruan J., Homer N., Marth G., Abecasis G., Durbin R. and 1000 Genome Project Data Processing Subgroup},
+  title     = {The Sequence Alignment/Map format and SAMtools},
+  journal   = {Bioinformatics},
+  year      = {2012},
+  volume    = {25},
+  number    = {16},
+  pages     = {2078-2079},
+  doi       = {10.1093/bioinformatics/btp352},
+}
+
+@Article{tarasov,
+  author    = {Tarasov A., Vilella A., Cuppen E., Nijman I. J. and Prins P.},
+  title     = {Sambamba: fast processing of NGS alignment formats},
+  journal   = {Bioinformatics},
+  year      = {2015},
+  volume    = {31},
+  number    = {12},
+  pages     = {2032-2034},
+  doi       = {10.1093/bioinformatics/btv098},
+}
+
